cleanups and optimisations
